Define non-zero freshness value for all docs
Consider this a conceptual PR and not tested code change. I can't build,
so I can't test.  I removed the pragma no-cache because that was
considered obsolete in 2005
(https://support.microsoft.com/en-us/kb/165150) I removed the Expires
header and replaced it with max-age because they do the same thing but
max-age is easier to specify.  For queries that return all documents I
set the max-age to an hour.  Anyone searching for a specific file will
use a non-empty q value and therefore the caching max-age will be zero.